fraudsters-management/requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-logging>=3.5.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
functions-framework>=3.5.0
requests>=2.31.0
flask>=3.0.0
gunicorn>=21.2.0
playwright>=1.40.0

